add playable buffer for readability split audioHelpers in files add support for double value in floatparameter (var is a double anyway..) fix audioDeviceOutNode vu meter type add rms channel listeners for computing perchannel rms only when needed
fix looper predelay

+++ src/LGML/Source/FloatParameter.cpp
@@ -21,7 +21,7 @@ FloatParameter::FloatParameter(const String & niceName, const String &descriptio
 
 void FloatParameter::setValueInternal(var _value)
 {
-    value = jlimit<float>(minimumValue, maximumValue, _value);
+    value = jlimit<double>(minimumValue, maximumValue, _value);
 }
 
 FloatSliderUI * FloatParameter::createSlider(FloatParameter * target)
